Project - tracking experiment
Project focused on tracking experiment. When we have a problem to solve through machine learning and we cannot decide which way will lead us to goal, then we need to experiment with number factors like number of  epochs, model selection, dataset size, here in this project we made a way to run  each case one by one and best model is selected and saved . This is what our project is all about.